+++ OnlySnarf/bin/config.py
@@ -15,7 +15,7 @@ def checkBothCreds():
 
 # checks Google creds access
 def checkGoogle():
-    print("Checking Google Creds")
+    print("Checking Google Creds (uploads)")
     if not os.path.exists(Settings.get_google_path()):
         print("Missing Google Creds")
         print()
@@ -25,19 +25,23 @@ def checkGoogle():
         print("Google Auth Successful")
     else: 
         print("Google Auth Failure")
-    print()
-    main()
 
 # checks OnlyFans login process
 def checkOnlyFans():
-    print("Checking Twitter Creds (OnlyFans Login)")
+    print("Checking OnlyFans Creds")
     if not os.path.exists(Settings.get_config_path()):
         print("Missing Config Path")
         return main()
     OnlySnarf.auth()
     OnlySnarf.exit()
-    print()
-    main()
+
+def checkTwitter():
+    print("Checking Twitter Creds")
+    if not os.path.exists(Settings.get_config_path()):
+        print("Missing Config Path")
+        return main()
+    OnlySnarf.auth()
+    OnlySnarf.exit()
 
 # function that creates the missing config
 def createConfig():
@@ -60,22 +64,34 @@ def createConfig():
         except Exception as e:
             print(e)
             main()
-    else: print("OnlySnarf Config Exists")
-    # print()
-    # main()
+    else:
+        print("OnlySnarf Config Exists")
+        return True
+    return False
 
 # provides instructions for creating or refreshing google creds
 def googleInstructions():
     print("[Google Instructions From README Go Here]")
-    print()
-    main()
 
 # creates the config then prompts for missing credentials
 def setupConfig():
-    createConfig()
-    updateConfig()
-    print()
-    main()
+    alreadyCreated = createConfig()
+    if not alreadyCreated:
+        updateConfig()
+
+# receives input for Google login
+def receiveGoogle():
+    data = {}
+    data['username'] = input('Google Username: ')
+    data['password'] = input('Google Password: ')
+    return data
+
+def receiveOnlyFans():
+    data = {}
+    data['email'] = input('OnlyFans Email: ')
+    data['username'] = input('OnlyFans Username: ')
+    data['password'] = input('OnlyFans Password: ')
+    return data
 
 # receives input for Twitter login
 def receiveTwitter():
@@ -90,8 +106,6 @@ def refreshAll():
     setupConfig()
     removeGoogle()
     googleInstructions()
-    print()
-    main()
 
 # removes config.conf
 def removeConfig():
@@ -102,8 +116,6 @@ def removeConfig():
         print("Removed Config")
     else:
         print("Error: Failed to Find Config")
-    print()
-    main()
 
 # removes google creds
 def removeGoogle():
@@ -114,11 +126,41 @@ def removeGoogle():
         print("Removed Google Creds")
     else:
         print("Error: Failed to Find Google Creds")
-    print()
-    main()
 
 # receives input for twitter login and saves to config.conf
 def updateConfig():
+    updateOnlyFans()
+    updateGoogle()
+    updateTwitter()
+
+def updateOnlyFans():
+    data = receiveOnlyFans()
+    # update conf variables username and password
+    # save the conf file
+    import fileinput
+    # Does a list of files, and
+    # redirects STDOUT to the file in question
+    for line in fileinput.input(Settings.get_config_path(), inplace = 1): 
+        line.replace("email None", "username {}".format(data['email']))
+        line.replace("username None", "username {}".format(data['username']))
+        line.replace("password None", "password {}".format(data['password']))
+        print(line)
+    print("OnlyFans Config Updated")
+
+def updateGoogle():
+    data = receiveGoogle()
+    # update conf variables username and password
+    # save the conf file
+    import fileinput
+    # Does a list of files, and
+    # redirects STDOUT to the file in question
+    for line in fileinput.input(Settings.get_config_path(), inplace = 1): 
+        line.replace("username_google None", "username {}".format(data['username']))
+        line.replace("password_google None", "password {}".format(data['password']))
+        print(line)
+    print("Google Config Updated")
+
+def updateTwitter():
     data = receiveTwitter()
     # update conf variables username and password
     # save the conf file
@@ -126,13 +168,11 @@ def updateConfig():
     # Does a list of files, and
     # redirects STDOUT to the file in question
     for line in fileinput.input(Settings.get_config_path(), inplace = 1): 
-        line.replace("username None", "username {}".format(data['username']))
-        line.replace("password None", "password {}".format(data['password']))
+        line.replace("username_twitter None", "username {}".format(data['username']))
+        line.replace("password_twitter None", "password {}".format(data['password']))
         print(line)
-    print("Config Updated")
-    print()
-    main()
-
+    print("Twitter Config Updated")
+    
 # this script is supposed to have menu options for 
 # ) creating the .conf file
 # ) updating the .conf file
@@ -140,23 +180,32 @@ def updateConfig():
 # ) a function for checking the google creds
 # when ran in it should check for the .conf file and google_creds
 def main():
-    createConfig()
-    return
-    print("-- Preparing OnlySnarf --")
+    print("-- OnlySnarf Config --")
     print("------------------------------")
     if os.path.isfile(Settings.get_config_path()):
         print(colorize("[*] Config File", 'conf')+": "+colorize("True", 'green'))
-        if str(Settings.get_username()) != "None":
-            print(colorize("[-] OnlyFans Username", 'conf')+": "+colorize(Settings.get_username(), 'green'))
+        if str(Settings.get_email()) != "None":
+            print(colorize("[-] OnlyFans Email", 'conf')+": "+colorize(Settings.get_email(), 'green'))
         else:
-            print(colorize("[-] OnlyFans Username", 'conf')+": "+colorize("", 'red'))
+            print(colorize("[-] OnlyFans Email", 'conf')+": "+colorize("", 'red'))
         if str(Settings.get_password()) != "None":
             print(colorize("[-] OnlyFans Password", 'conf')+": "+colorize("******", 'green'))
         else:
             print(colorize("[-] OnlyFans Password", 'conf')+": "+colorize("", 'red'))
-
+        if str(Settings.get_username()) != "None":
+            print(colorize("[-] OnlyFans Username", 'conf')+": "+colorize(Settings.get_username(), 'green'))
+        else:
+            print(colorize("[-] OnlyFans Username", 'conf')+": "+colorize("", 'red'))
+        if str(Settings.get_username_google()) != "None":
+            print(colorize("[-] Google Username", 'conf')+": "+colorize(Settings.get_username_google(), 'green'))
+        else:
+            print(colorize("[-] Google Username", 'conf')+": "+colorize("", 'red'))
+        if str(Settings.get_password_google()) != "None":
+            print(colorize("[-] Google Password", 'conf')+": "+colorize("******", 'green'))
+        else:
+            print(colorize("[-] Google Password", 'conf')+": "+colorize("", 'red'))
         if str(Settings.get_username_twitter()) != "None":
-            print(colorize("[-] Twitter Username", 'conf')+": "+colorize(Settings.get_username_twitter(), 'green'))
+            print(colorize("[-] Twitter Username", 'conf')+": "+colorize(Settings.get_username_google(), 'green'))
         else:
             print(colorize("[-] Twitter Username", 'conf')+": "+colorize("", 'red'))
         if str(Settings.get_password_twitter()) != "None":
@@ -171,39 +220,43 @@ def main():
         print(colorize("[*] Google Creds", 'conf')+": "+colorize("False", 'red'))
     print("------------------------------")
     print(colorize("Menu:", 'menu'))
-    print(colorize("[ 0 ]", 'menu') + " Check Credentials - Google")
-    print(colorize("[ 1 ]", 'menu') + " Check Credentials - Twitter")
-    # print(colorize("[ 2 ]", 'menu') + " Check Credentials - Both")
-    print(colorize("[ 2 ]", 'menu') + " Config - Create")
-    print(colorize("[ 3 ]", 'menu') + " Config - Update")
+    print(colorize("[ 0 ]", 'menu') + " Config - Create")
+    print(colorize("[ 1 ]", 'menu') + " Config - Update - Google")
+    print(colorize("[ 2 ]", 'menu') + " Config - Update - OnlyFans")
+    print(colorize("[ 3 ]", 'menu') + " Config - Update - Twitter")
     print(colorize("[ 4 ]", 'menu') + " Config - Remove")
-    print(colorize("[ 5 ]", 'menu') + " Google Creds - Instructions")
-    print(colorize("[ 6 ]", 'menu') + " Google Creds - Remove")
+    print(colorize("[ 5 ]", 'menu') + " Google Creds - Check")
+    print(colorize("[ 6 ]", 'menu') + " Google Creds - Instructions")
+    print(colorize("[ 7 ]", 'menu') + " Google Creds - Remove")
     # print(colorize("[ 8 ]", 'menu') + " Refresh All")
     while True:
         choice = input(">> ")
         try:
             if int(choice) < 0 or int(choice) >= 9: raise ValueError
-            if int(choice) == 0:
-                checkGoogle()
-            elif int(choice) == 1:
-                checkOnlyFans()
             # elif int(choice) == 2:
             #     checkBothCreds()
-            elif int(choice) == 2:
+            if int(choice) == 0:
                 setupConfig()
+            elif int(choice) == 1:
+                updateGoogle()
+            elif int(choice) == 2:
+                updateOnlyFans()
             elif int(choice) == 3:
-                updateConfig()
+                updateTwitter()
             elif int(choice) == 4:
                 removeConfig()
             elif int(choice) == 5:
-                googleInstructions()
+                checkGoogle()
             elif int(choice) == 6:
+                googleInstructions()
+            elif int(choice) == 7:
                 removeGoogle()
             # elif int(choice) == 8:
             #     refreshAll()
         except (ValueError, IndexError, KeyboardInterrupt):
             print("Error: Incorrect Index")
+    print()
+    main()
 
 ###########################